# Conflicts:
#	src/hfe.c

+++ src/hfe.c
@@ -890,8 +890,8 @@ static void hfe_seek(int drive, int track)
   hfe_info[drive]->track_data = trackbits;
   hfe_info[drive]->track_data_bytes = track_len;
   hfe_info[drive]->poll_calls_per_bit = encoding ? 1 : 2;
-  log_debug("hfe: seek: loaded %zu bytes of data for drive %d track %d at %p",
-            hfe_info[drive]->track_data_bytes,
+  log_debug("hfe: seek: loaded %lu bytes of data for drive %d track %d at %p",
+            (unsigned long)hfe_info[drive]->track_data_bytes,
             drive,
             track,
             hfe_info[drive]->track_data);
